// src/app/model/parts-search-model.ts
export class PartsSearchModel {
  number?: string;
  id?: number;
  machineProductId?: number;
  name?: string;
  description?: string;
  partType?: string;
  make?: string;
  productLine?: string;
  subProductLine?: string;
  skuName?: string;
  keyword?: string;
  lineOfService?: string;
  searchString?: string;
  active?: boolean;
  requiresAudit?: boolean;
  pageSize?: number;
  pageNumber?: number;
  partCategory?: string;
  isActionProduct?: boolean;
}